+++ CHANGELOG.md
@@ -1,15 +1,24 @@
+## 0.0.5
+
+- Add media query size extension base on Fluent 2 layout rule 
+- Add Fluent 2 Layout Widget base on Fluent 2 layout rule
+- Add media query size extension base on Material 3 layout rule 
+- Add Material 3 Layout Widget base on Material 3 layout rule
+- Update Example
+- Update README.md
+
 ## 0.0.4
 
-- add verticalMargin
-- add horizontalMargin
+- Add verticalMargin
+- Add horizontalMargin
 
 ## 0.0.3
 
-- change baseMargin() - basePadding() becom public, update example
+- Change baseMargin() - basePadding() becom public, update example
 
 ## 0.0.2
 
-- update README.md , update example
+- Update README.md , update example
 
 ## 0.0.1